Add fence after volatile field stores
We override the various store evaluators in OpenJ9 and emit a fence
following volatile stores to adhere to the Java Memory Model. This is
necessary for functional correctness.

There are two cases to handle:

1. Field is resolved
2. Field is unresolved

For 1. the implementation is trivial in that we simply emit the fence
after the store evaluator has emitted the necessary instructions to
evaluate the store node.

For 2. the implementation is more complicated. We do not know at
compile time whether the field in question is volatile or not. To
combat this we emit a NOP placeholder and in the runtime resolution
logic we call a VM helper to determine whether the resolved field is
volatile or not. If we determine the field is volatile we overwrite
the NOP in the mainline instruction stream with a fence.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>