Add Unsafe.copyMemory to dontInline list
In JDK11, sun/misc/Unsafe methods works as Java wrappers to internal
implementation of Unsafe in package jdk.internal.misc. So in JDK11,
call to sun/misc/Unsafe.copyMemory contains a call chain to
jdk.internal.misc.Unsafe.copyMemory which later calls JNI method.
In JDK8, we do not inline this method and let GlobalValuePropagation to
perform optimization where it convert call to System.arrayCopy calls
which all platforms later inlines using optimized instruction sequence.
Due to different call chain in JDK11, we inline the method and resulting
tree, restricts the GlobalValuePropagation leaving final method with
JNI call which is slower. As GlobalValuePropagation is going to identify
the method and optimize it, this commit adds the
sun/misc/Unsafe.copyMemory to the list recognized methods which we do
not inline.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>