Extend the source register in ArrayStoreCHK evaluator
In case of compressed refs configuration in ArrayStoreCHK, we get the
originial decompressed node from by traversing to the second source
child of wbBarrier node. But in case if the source child contains known
NULL object, we would end up with a source child node of type TR::Int32
which when we evaluate, will only write into lower half of the register.
This could lead to scenarios where we have garbage value in the upper
half of the register which could lead us to dereference garbage.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>