Get first slot parameter for non-static method call from resolvedMethodSymbol
While inlining non-static method calls, we build the parameters where
the very first slot of the parameters is the class of which that method
belongs to. We later build ArgInfo from the method symbol of the callee
by iterating through this list of parameters and for any parameter which
is Class, we query VM to get the class pointer from the signature of
the class. We do not need to call VM to get the class pointer for the
first parameter slot of the non-static method calls and we have that
information avaible through resolvedMethodSymbol.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>