Change Query to check VFE1 feature for FMA
While checking if we can use Vector FMA instruction in Math.FMA
recognition, we used to check the flags on CPU to see if
VectorEnhancementFacility 1 is installed or not. These flags are
obsolete and we should be checking feature. This commit change the query
to check feature on CPU instead of flags.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>