Update the getRawCount dump information without compiler object
When dumping the block frequency information current implementation
simply prints the value stored in the the internal frequency array which
may not reflect the correct frequency of the byte code info, it needs to
compute the correct frequency using getRawCount function and print the
information. Currently there was a limitation in getRawCount function
where we do not expect to call that using NULL callSiteInfo or NULL
compilation object, which is not necessary in order to dump the
information. This puts a safeguard in place so that getRawCount can be
used with NULL callSiteInfo and compiler object.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>