Increase the cutoff to inline interpreted targets
In hot and above compilatios, if a target it is looking to inline is still
interpreted, increase the frequency cut-off for such targets. Currently this is
not extended to warm compiles, but changes in this commit also adds environment
variable to apply higher frequency cut-off for interpreted targets in warm
compilations.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>